Abstract role options dropdown used in 3 places more globally, update generic map type to clean up typescript
