+++ frontend/pages/admin/UserManagementPage/components/UserForm/UserForm.tsx
@@ -22,6 +22,7 @@ import InfoBanner from "components/InfoBanner/InfoBanner";
 import CustomLink from "components/CustomLink";
 import SelectedTeamsForm from "../SelectedTeamsForm/SelectedTeamsForm";
 import SelectRoleForm from "../SelectRoleForm/SelectRoleForm";
+import { roleOptions } from "../../helpers/userManagementHelpers";
 
 const baseClass = "create-user-form";
 
@@ -34,53 +35,6 @@ enum UserTeamType {
   GlobalUser = "GLOBAL_USER",
   AssignTeams = "ASSIGN_TEAMS",
 }
-
-interface GlobalRoleParams {
-  isPremiumTier: boolean;
-  isApiOnly?: boolean;
-}
-
-// Create and edit global user role
-const globalUserRoles = ({
-  isPremiumTier,
-  isApiOnly,
-}: GlobalRoleParams): IRole[] => {
-  const roles: IRole[] = [
-    {
-      disabled: false,
-      label: "Observer",
-      value: "observer",
-    },
-    {
-      disabled: false,
-      label: "Maintainer",
-      value: "maintainer",
-    },
-    {
-      disabled: false,
-      label: "Admin",
-      value: "admin",
-    },
-  ];
-
-  if (isPremiumTier) {
-    roles.unshift({
-      disabled: false,
-      label: "Observer+",
-      value: "observer_plus",
-    });
-    // Next release
-    // if (isApiOnly) {
-    //   roles.splice(3, 0, {
-    //     disabled: false,
-    //     label: "GitOps",
-    //     value: "gitops",
-    //   });
-    // }
-  }
-
-  return roles;
-};
 
 export interface IFormData {
   email: string;
@@ -339,7 +293,7 @@ const UserForm = ({
           label="Role"
           value={formData.global_role || "Observer"}
           className={`${baseClass}__global-role-dropdown`}
-          options={globalUserRoles({
+          options={roleOptions({
             isPremiumTier,
             isApiOnly,
           })}
